Initial working version
A working version of the program inducer.
It features function invention and reuse (although limited), type based
pruning and searching for a program using iterative deepening.
Currently the initial program and the bk knowledge need to be hardcoded
in the file `Interpreter.hs`. The next version should fix all the
hardcoding.